Move className generation to css-in-js
