+++ packages/css-in-js/.eslintrc.js
@@ -9,6 +9,7 @@ module.exports = {
       'newlines-between': 'always',
       groups: ['builtin', 'external', 'internal', ['parent', 'sibling', 'index']]
     }],
+    'no-extra-parens': 'off'
   },
   'env': {
     'jest': true,
