+++ packages/css-in-js/src/index.js
@@ -1,6 +1,7 @@
 // @flow
 
-export * from './tags';
-export * from './sheet';
-export * from './transpile';
 export * from './selectors';
+export * from './sheet';
+export * from './tags';
+export * from './toClassName';
+export * from './transpile';
